Fix clang 15 -Wstrict-prototype warnings
Clang 15 has become more strict about prototypes of functions without arguments in C, showing many warnings similar to:

    src/cbor/maps.h:49:49: warning: a function declaration without a prototype is deprecated in all versions of C [-Wstrict-prototypes]
    CBOR_EXPORT cbor_item_t *cbor_new_indefinite_map();
                                                    ^
                                                     void

This is because in C, functions declared or defined with () are considered old-style functions with variadic arguments. (Although these will likely be finally removed in C23.)

Fix the warnings by using `(void)` for each of the declarations and definitions. This also fixes two incorrect invocations of `cbor_new_indefinite_map(void)` in the tests.